+++ addons/sapian_core/tests/__init__.py
@@ -1,2 +1,3 @@
 # -*- coding: utf-8 -*-
 from . import test_onboarding_wizard
+from . import test_onboarding_web